requirements.txt
aiohttp
alembic==1.0.0
Django==2.1
djangorestframework==3.8.2
edgedb
nltk==3.3
psycopg2-binary==2.7.7
pymongo
progress
uvloop
